geisa-imx93-common: add ssh server in production image
Ssh server was only available in the debug image, but it is
needed for the geisa conformance tests. Then add the ssh server
to the common image, so it will be available in both production
and debug images.
To allow ssh server to accept root login, the image feature
`allow-root-login` is added.

SFL: #17867
Change-Id: Iaa5fbce6f093f771501622f81e662f4b535573fb