systemd: add support for setting RuntimeWatchdogSec
Systemd recipe is only configuring the RebootWatchdogSec option which
enables and configures the watchdog only when a reboot is requested.
As it is also possible that the system hangs during runtime, allow
configuring (and therefore enabling) the RuntimeWatchdogSec option.

By default, WATCHDOG_TIMEOUT is set to 60, then the watchdog is now
enabled and triggered by systemd.

SFL: #17871
Change-Id: I9e211b99cf90cee4ece8ad7ae03a67132ef30861